gemini-vision/requirements.txt
streamlit
google-adk
python-dotenv
